fix compile error

+++ apps/siot-net/src/main.c
@@ -24,7 +24,7 @@ static const struct nvs_point nvs_pts[] = {
 
 static struct net_mgmt_event_callback mgmt_cb;
 
-static void net_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void net_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 			      struct net_if *iface)
 {
 	if (mgmt_event == NET_EVENT_L4_CONNECTED) {
